+++ tools/build-single-file.mjs
@@ -13,6 +13,7 @@ const css_files = [
   'src/styles/editor.css',
   'src/styles/results.css',
   'src/styles/dashboard.css',
+  'src/styles/schema.css',
   'src/styles/vendors/uplot.css',
 ];
 
@@ -30,6 +31,7 @@ const script_files = [
   'src/scripts/query-results/charts.js',
   'src/scripts/query-editor/lexer.js',
   'src/scripts/dashboard/dashboard.js',
+  'src/scripts/schema/schema.js',
   'src/scripts/boot.js',
   'src/scripts/query-results/download.js',
 ];